+++ CHANGELOG.md
@@ -57,6 +57,9 @@ in the class `Update`. The bot must be an administrator in the chat to receive t
 - The class `GiveawayWinners` and the property `GiveawayWinners` to the class `Message` for messages about the completion of a giveaway with public winners.
 - The class `GiveawayCompleted` and the property `GiveawayCompleted` to the class `Message` for service messages about the completion of a giveaway without public winners.
 - New `MessageType` enum members: `Giveaway`, `GiveawayCreated`, `GiveawayWinners` and `GiveawayCompleted`
+- The properties `AccentColorId`, `BackgroundCustomEmojiId`, `ProfileAccentColorId`, and `ProfileBackgroundCustomEmojiId` to the class `Chat`.
+- The property `HasVisibleHistory` to the class `Chat`.
+- Classes `MaybeInaccessibleMessage` and `InaccessibleMessage`.
 
 ### Changed
 
@@ -107,6 +110,10 @@ in the class `Update`. The bot must be an administrator in the chat to receive t
 - Renamed the class `UserShared` to `UsersShared` and changed the property `UserId` to `UserIds`.
 - Replaced the property `UserShared` in the class Message with the property `UsersShared`.
 - Replaced enum member `MessageType.UserShared` with `MessageType.UsersShared`
+- The class `MessageOrigin` and replaced the fields `ForwardFrom`, `ForwardFromChat`, `ForwardFromMessageId`, `ForwardSignature`, `ForwardSenderName` 
+and `ForwardDate` with the field `ForwardOrigin` of type `MessageOrigin in the class `Message`.
+- Type of the property `Message` of the class `CallbackQuery` to `MaybeInaccessibleMessage`
+- Type of the property `PinnedMessage` of the class `Message` to `MaybeInaccessibleMessage`.
 
 ## [v20.0.0] - Unreleased